requirements.txt
attrs==25.3.0 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.234 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-cloud-assembly-schema==41.2.0 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-lib==2.194.0 ; python_version >= "3.9" and python_version < "4.0"
cattrs==24.1.3 ; python_version >= "3.9" and python_version < "4.0"
constructs==10.4.2 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.9" and python_version < "3.11"
func-args==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-resources==6.5.2 ; python_version >= "3.9" and python_version < "4.0"
jsii==1.111.0 ; python_version >= "3.9" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.13.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.21.0 ; python_version == "3.9"
